Update al local storage
Se guarda y se refresca la página para que se muestren los botones correctos en el navbar

+++ compawnion-app/src/components/Navbar.js
@@ -1,4 +1,4 @@
-import React, { Fragment } from "react";
+import React, { Fragment, useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import { Button, AppBar, Toolbar, Typography, IconButton } from "@mui/material";
 import { makeStyles } from "@mui/styles";
@@ -9,12 +9,20 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 const handleClick = (e) => {
-  localStorage.clear();
+  var answer = window.confirm("¿Desea cerrar sesión?");
+
+  if (answer) {
+    localStorage.clear();
+    Window.location.reload();
+  }
 };
 
 export default function Navbar() {
+  const localUser = JSON.parse(localStorage.getItem("usuario"));
+
+  const [userData, setUserData] = useState(localUser);
+
   const classes = useStyles();
-  const data = localStorage.getItem("usuario");
 
   return (
     <Fragment>
@@ -35,43 +43,75 @@ export default function Navbar() {
           >
             COMPAWNION
           </Typography>
-          <IconButton
-            aria-label="mascota"
-            component={Link}
-            color="inherit"
-            to="/publicar"
-            sx={{ marginLeft: "auto" }}
-          >
-            <AddCircle />
-          </IconButton>
-          <Button
-            component={Link}
-            to="/perfil/:id"
-            color="inherit"
-            size="large"
-          >
-            Perfil
-          </Button>
-          <Button component={Link} to="/follow-up" color="inherit" size="large">
-            Follow Up
-          </Button>
-          <Button component={Link} to="/busqueda" color="inherit" size="large">
-            Buscar
-          </Button>
-          {data !== null ? (
-            <Button
-              component={Link}
-              to="/"
-              color="inherit"
-              size="large"
-              onClick={handleClick}
-            >
-              Log Out
-            </Button>
+
+          {userData !== null ? (
+            <>
+              <IconButton
+                aria-label="mascota"
+                component={Link}
+                color="inherit"
+                to="/publicar"
+                sx={{ marginLeft: "auto" }}
+              >
+                <AddCircle />
+              </IconButton>
+              <Button
+                component={Link}
+                to="/perfil/:id"
+                color="inherit"
+                size="large"
+              >
+                Perfil
+              </Button>
+              <Button
+                component={Link}
+                to="/follow-up"
+                color="inherit"
+                size="large"
+              >
+                Follow Up
+              </Button>
+              <Button
+                component={Link}
+                to="/busqueda"
+                color="inherit"
+                size="large"
+              >
+                Buscar
+              </Button>
+              <Button
+                component={Link}
+                to="/"
+                color="inherit"
+                size="large"
+                onClick={handleClick}
+              >
+                Log Out
+              </Button>
+            </>
           ) : (
-            <Button component={Link} to="/login" color="inherit" size="large">
-              Log In
-            </Button>
+            <>
+              <Button
+                component={Link}
+                to="/follow-up"
+                color="inherit"
+                size="large"
+                sx={{ marginLeft: "auto" }}
+              >
+                Follow Up
+              </Button>
+              <Button
+                component={Link}
+                to="/busqueda"
+                color="inherit"
+                size="large"
+              >
+                Buscar
+              </Button>
+              <Button component={Link} to="/login" color="inherit" size="large">
+                Log In
+              </Button>
+            </>
           )}
         </Toolbar>
       </AppBar>
